+++ lib/decorator.rb
