+++ lib/teacher.rb
