+++ lib/trimmer.rb
